Default behaviour is to set soap:action to the default wsa:Action which follows the syntax namespace/portType.name/operation.name/input.name
Can be overridden by specifying an optional property "blankSOAPAction" with value true. 

This is to overcome an interoperability issue between WAS and jBoss. A WAS consumer when using WS-Addresssing
sets a default value in SOAPAction if SOAPAction is blank "" in the WSDL. A jBoss provider fails when the request 
contains the WAS default value while the WSDL contains a blank value "".